requirements.txt
streamlit
langchain
langchain-openai
openai
faiss-cpu
python-dotenv
gtts
